requirements.txt
streamlit
google-genai
google-cloud-firestore
streamlit-autorefresh